requirements.txt
# Runing
flask
flask_sqlalchemy
psycopg2-binary
python-dotenv

# Testing
pytest
pytest-mock
flask-testing
requests
pytest-cov
factory_boy        # Create mock data objects easily
Faker              # Generate random names, books, etc.
